# docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Template Swagger Docs
  description: This API is used for Template Service
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/LieAlbertTriAdrian/todos-service/template/1.0.0
  - url: http://template.swagger.io/v1
paths:
  /healthcheck:
    get:
      tags:
        - healthcheck
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Healthcheck'
  /todos:
    post:
      tags: 
        - todos
      operationId: createTodo
      summary: 'Create a Todo item'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '200':
          description: 'Todo created successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    get:
      tags: 
        - todos
      operationId: fetchTodos
      summary: 'Fetch Todo list items'
      parameters: 
        - name: cursor
          in: query
          schema:
            type: string
          description: 'Cursor used for pagination to fetch the next pages'
        - name: limit
          in: query
          schema:
            type: integer
          description: 'Limit used for to limit the number of item to be fetched'
      responses:
        '200':
          description: 'Todo list fetched successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
components:
  schemas:
    Healthcheck:
      type: object
      properties:
        status:
          type: string
          example: 'OK'
        dependencies:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: 'postgresql'
              status:
                type: string
                example: 'OK'
    Todo:
      properties:
        id:
          type: string
          readOnly: true
          example: 'abcd-3254fsd-3242'
        status:
          type: string
          example: 'IN-PROGRESS'
        text:
          type: string
          example: 'Feeding the Hamster'
        createdAt:
          type: string
          format: date-time